@@ -1,7 +1,7 @@
 [package]
 name = "lgi_schedule"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -9,3 +9,4 @@ edition = "2021"
 #rand = "0.8.5"
 jiff = { version = "0.2.0", features = ["js", "serde"] }
 serde = { version = "1.0.219", features = ["derive"] }
+quick-xml = "0.38"
